Laid groundwork for Impact spreading. More to do. Need to find a way to copy dots directly, since it copies dots with remaining time even (spread an ignite with 1s left, you get multiple ignites with 1s left)
Also fixed base hot streak proc to account for the fact that it scales with your crit in a linear fashion. See discord #cata-mage for more info.

Also fixed combustion in a couple ways: 1) Combustion tick crits do not proc ignite. 2) Combustion double-dips on mastery for living bomb and pyroblast but not ignite