hunter: fix survival aoe apl
The previous AOE APL had many problems, the largest of which was being
in Aspect of the Fox, leading to essentially infinite focus, presumably
due to the targets attacking the player.

The new APL is fixes that and makes some improvements:

- The result with no condition for Explosive Trap seems to always be
better.
- Serpent Sting can be ignored since it will be applied passively via
Multi Shot, which is a high priority cast.
- Mutli Shot seems to be worth spamming on aoe regardless of target count.
- The Kill Shot wait seems to not matter.
- Explosive Shot should only be used with Lock and Load. The same
  condition as the single target APL is used, since it's equivalent in
  dps to only using Explosive Shot when the debuff is not present on the
  target.

One notable thing still missing is Trap Launcher.