core: fix ghost attacks when pets are disabled during prepull
During prepull inside pet.Enable(), we enable auto attacks with a
pending action. However, if the pet is disabled again before this action
can run we end up in a situation where the pet unit is disabled, but is
stuck in a bugged state where it's still auto attacking and using
abilities, while auras and such are not removed.

To guard against this, always check if the unit is actually enabled
before enabling auto attacks.

Fixes #518