core: fix dot.NextTickAt()
Calculating the next tick time by taking the last tick time and adding
the tick period produces incorrect results when the dot is refreshed in
between with a different haste amount.

So instead we return NextActionAt of the pending action responsible for
ticks in the first place, which should always be accurate.

Since NextTickAt() was used in the calculation of the aura duration for
refreshing dots, this also fixes issues where the aura duration doesn't
match the time of the last tick.

That part seems to only affect fire mage damage, since they are the only
ones that have a hasted dot effect that is both sometimes refreshed and
generally not renewed (Pyroblast dot). With warlock dots this doesn't
happen because the test APLs always refresh dots before they expire.

Warrior damage is also affected since they used NextTickAt() directly
for rend double ticks.