core: fix NumTicksRemaing() miscounting on 0 delta
When we just had a tick, lastTickTime is just sim.CurrentTime which
makes the total expression we return:

```
(sim.CurrentTime - sim.CurrentTime +
  dot.tickPeriod*time.Duration(maxTicksRemaining))/dot.tickPeriod + 1
```

Which simplifies to just `maxTicksRemaining + 1`, which is wrong. If
lastTickTime is just 1ns different on the other hand, we get a value
that is correctly truncated down, and adding the 1 results in the
correct amount of ticks.