core: completely rework the dot code
Removes all the spaghetti code inside the dot core with all it's
confusing behaviour and bugs and replaces it with a much saner base.
There's now a much clearer boundry between what's considered "inside"
and "outside" of the dot code and it's responsibilities which should
make it much less brittle.

This should also improve performance significantly by not recreating
periodic actions every time we refresh a dot, as well as removing
redundant calculations.