core: don't reschedule tick on Apply()
Apply() does not actually change the next tick timer at all, but the old
code still recreated it every time for some reason. This commit changes
that.

This causes tests to change as the order of some actions that happen at
the same time is now different, because they are not added to the "end"
of the queue for that moment in time by AddPendingAction() anymore.